src/requirements.txt
pillow>=10.3,<11